dom/base/test/jsmodules/importmaps/multiple/test_importMap_after_module_load.html

<!DOCTYPE html>
<head>
  <meta charset=utf-8>
  <title>Test dynamic imports before and after an import map</title>
</head>
<body>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script>
  SimpleTest.waitForExplicitFinish();

  // eslint-disable-next-line no-unused-vars
  let gotMsg = false;
  let console = Cc["@mozilla.org/consoleservice;1"].getService(Ci.nsIConsoleService);
  function gotMsgPromise() {
    return new Promise(resolve => {
      let listener = {
        QueryInterface: ChromeUtils.generateQI(["nsIConsoleListener"]),
        observe(msg) {
          info("console message:" + msg);
          ok(msg.logLevel == Ci.nsIConsoleMessage.warn, "log level should be 'warn'.");
          console.unregisterListener(this);
          resolve();
          gotMsg = true;
        }
      };
      console.registerListener(listener);
    });
  }

  let hasError = false;
  // eslint-disable-next-line no-unused-vars
  function importMapError() {
    hasError = true;
  }

  // result will be set in module_simpleImport.mjs
  var result;

  // This is a module load before the import map tag.
  // It will import "./module_simpleExport.mjs", which makes "module_simpleExport.mjs"
  // as a resolved module.
  import("./module_simpleImport.mjs").then(() => {
    ok(result == 42, 'Check the imported value result: ' + result);
    let script = document.createElement("script");
    script.type = "importmap";
    script.addEventListener("error", () => importMapError());
    script.textContent = `{
      "imports": {
        "./module_simpleExport.mjs": "./scope1/module_simpleExport.mjs"
      }
    }`;

    // The console warning message is triggered during appendChild(script),
    // so register the console listener before appending the script.
    let p = gotMsgPromise();
    document.head.appendChild(script);
    return p;
  }).then(() => {
    ok(gotMsg, "Should have got the console warning.");
    return import("./module_simpleExport.mjs");
  }).then((ns) => {
    ok(ns.x == 42, 'Check simple imported value ns.x: ' + ns.x);
    ok(!hasError, "onerror of the import map shouldn't be called.");
  }).catch((e) => {
    ok(false, "throws " + e);
  }).then(() => {
    SimpleTest.finish();
  });
</script>
</body>
